Update mame.h
markwkidd ported nayslayer's vector fix from mame2003

+++ src/mame.h
@@ -232,7 +232,7 @@ struct GameOptions
       
   int		   vector_width;	       /* requested width for vector games; 0 means default (640) */
   int		   vector_height;	       /* requested height for vector games; 0 means default (480) */
-  int		   beam;			           /* vector beam width */
+  float    beam;                 /* vector beam width */
   int	     vector_flicker;	     /* vector beam flicker effect control */
   float	   vector_intensity_correction;   
   int		   translucency;	       /* 1 to enable translucency on vectors */
